[DEV-757] moved masternode to integration directory
