+++ src/app.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AinModule } from './blockchain/ain/ain.module';
 import { BlockchainModule } from './blockchain/blockchain.module';
 import { GetConfig } from './config/config';
+import { IntegrationModule } from './integration/integration.module';
 import { SharedModule } from './shared/shared.module';
 import { AnalyticsModule } from './subdomains/analytics/analytics.module';
 import { StakingModule } from './subdomains/staking/staking.module';
@@ -14,11 +14,11 @@ import { UserModule } from './subdomains/user/user.module';
   imports: [
     TypeOrmModule.forRoot(GetConfig().database),
     SharedModule,
-    AinModule,
     UserModule,
     StakingModule,
     AnalyticsModule,
     BlockchainModule,
+    IntegrationModule,
   ],
   controllers: [AppController],
   providers: [AppService],
