+++ src/integration/masternode/application/dto/create-masternode.dto.ts
