+++ src/integration/masternode/application/dto/prepare-resign-masternode.dto.ts
