+++ src/integration/masternode/application/repositories/masternode.repository.ts
