+++ src/integration/masternode/application/services/masternode.service.ts
@@ -6,7 +6,7 @@ import { HttpError, HttpService } from 'src/shared/services/http.service';
 import { SettingService } from 'src/shared/services/setting.service';
 import { IsNull, LessThan, MoreThan, Not } from 'typeorm';
 import { Masternode } from '../../domain/entities/masternode.entity';
-import { MasternodeState } from '../../../../domain/enums';
+import { MasternodeState } from '../../../../subdomains/staking/domain/enums';
 import { ResignMasternodeDto } from '../dto/resign-masternode.dto';
 import { MasternodeRepository } from '../repositories/masternode.repository';
 import { CreateMasternodeDto } from '../dto/create-masternode.dto';
